refactor: centralize SIREN parsing and consolidate file upload logic
- Move `parseSiren` to shared module for reusability across packages
- Refactor `useFileUploadForm` to accept `saveMutation` and handle upload internally
- Remove duplicate upload logic from Step2Upload and JointEvaluationForm
- Simplify button disabled state by exporting `isPending` from the hook
