+++ packages/app/src/modules/my-space/index.ts
@@ -1,4 +1,4 @@
 export { CompanyDeclarationsPage } from "./CompanyDeclarationsPage";
-export { extractSiren, formatSiren, parseSiren } from "./formatSiren";
+export { extractSiren, formatSiren } from "./formatSiren";
 export { MonEspacePage } from "./MonEspacePage";
 export { MyCompaniesPage } from "./MyCompaniesPage";
