+++ packages/app/src/modules/shared/useFileUploadForm.ts
@@ -3,15 +3,22 @@
 import { useCallback, useRef, useState } from "react";
 
 import { getDsfrModal } from "./getDsfrModal";
+import { uploadFile } from "./uploadFile";
 
-type Options = {
-	onConfirm: () => void;
+type SaveMutation = {
+	mutate: (input: { fileName: string; filePath: string }) => void;
+	isPending: boolean;
 };
 
-export function useFileUploadForm({ onConfirm }: Options) {
+type Options = {
+	saveMutation: SaveMutation;
+};
+
+export function useFileUploadForm({ saveMutation }: Options) {
 	const modalRef = useRef<HTMLDialogElement>(null);
 	const [selectedFile, setSelectedFile] = useState<File | null>(null);
 	const [uploadError, setUploadError] = useState<string | null>(null);
+	const [isUploading, setIsUploading] = useState(false);
 
 	function handleFileChange(file: File | null, error: string | null) {
 		setSelectedFile(file);
@@ -50,9 +57,25 @@ export function useFileUploadForm({ onConfirm }: Options) {
 		openModal();
 	}
 
-	function handleConfirm() {
+	async function handleConfirm() {
 		closeModal();
-		onConfirm();
+		if (!selectedFile) return;
+
+		setIsUploading(true);
+		try {
+			const result = await uploadFile(selectedFile);
+			if (!result.ok) {
+				setUploadError(result.error);
+				setIsUploading(false);
+				return;
+			}
+			saveMutation.mutate({
+				fileName: selectedFile.name,
+				filePath: result.key,
+			});
+		} catch {
+			setIsUploading(false);
+		}
 	}
 
 	return {
@@ -60,6 +83,7 @@ export function useFileUploadForm({ onConfirm }: Options) {
 		handleConfirm,
 		handleFileChange,
 		handleSubmit,
+		isPending: isUploading || saveMutation.isPending,
 		modalRef,
 		selectedFile,
 		uploadError,
